Hold the timetable in a third of the memory
stop_times.txt is the largest thing this server keeps: 3.17 million rows
for the region, measured at 365 MB of live heap. Most of that was waste:

- each stop_id was a slice of its CSV record, and encoding/csv backs a
  record's fields with one string, so every row kept its whole line alive
- rows were grown per trip through append, keeping the doubling slack
- times and sequence were 64-bit ints

The file is now read twice: once to count rows per trip, once to fill a
single exactly-sized backing array. Stop ids are copied out and interned
(19,706 unique), trip ids cloned, and times and sequence are int32.
Measured on the live regional archive: 113 MB, and every one of the
93,026 trips is identical row for row to a plain parse of the file.

gtfsTimeString accepts either int width, so its many callers are
unchanged; the few struct literals and arithmetic sites convert.

+++ elk.go
@@ -160,10 +160,10 @@ func (r *elkRegion) loadStopTimesRaw() map[string][]StopTimeInfo {
 			seq, _ = strconv.Atoi(rec["stop_sequence"])
 		}
 		st[tid] = append(st[tid], StopTimeInfo{
-			arrival_time:   gtfsSeconds(rec["arrival_time"]),
-			departure_time: gtfsSeconds(rec["departure_time"]),
+			arrival_time:   int32(gtfsSeconds(rec["arrival_time"])),
+			departure_time: int32(gtfsSeconds(rec["departure_time"])),
 			stop_id:        rec["stop_id"],
-			stop_sequence:  seq,
+			stop_sequence:  int32(seq),
 		})
 	})
 	if len(st) == 0 {
@@ -565,7 +565,7 @@ func (r *elkRegion) enrich(payload []byte) []byte {
 				if stopID, _ := vehicle["stopId"].(string); stopID != "" {
 					for _, st := range r.stopTimesForTrip(tripID) {
 						if st.stop_id == stopID {
-							trip["delay"] = nowSec - st.departure_time
+							trip["delay"] = nowSec - int(st.departure_time)
 							break
 						}
 					}
